+++ src/AssemblyVersion.cs
@@ -5,5 +5,5 @@
   
  using System.Reflection;
 
- [assembly: AssemblyVersion("2.4.5.2")]
- [assembly: AssemblyFileVersion("2.4.5.2")]
+ [assembly: AssemblyVersion("2.4.5.3")]
+ [assembly: AssemblyFileVersion("2.4.5.3")]